requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==5.0.2
django-admin-interface==0.27.0
django-colorfield==0.11.0
django-crispy-forms==1.14.0
django-import-export==2.9.0
django-js-asset==2.0.0
django-tinymce==3.5.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
python-slugify==8.0.1
pytz==2022.4
PyYAML==6.0
sqlparse==0.4.3
tablib==3.2.1
text-unidecode==1.3
typing_extensions==4.3.0
xlrd==2.0.1
xlwt==1.3.0
